Use the sound player from its own file
The sound player class is to supposed to be shared between the
hack-toy-apps and the clubhouse projects, and thus, having it in its
file makes updating it easier.

Since the class is, however, called HackSoundServer in the hack-toy-apps
repo, this commit keeps the compatibility of using the SoundPlayer name
in the Clubhouse and importing it from the "system" module, for keeping
it simpler to use in the quests.

https://phabricator.endlessm.com/T24529
